Move v1 api to its own namespace
