+++ a-plus/urls.py
@@ -15,7 +15,7 @@ urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
     url(r'^shibboleth/', include(shibboleth_login.urls)),
     url('', include(social.apps.django_app.urls, namespace='social')),
-    url(r'^api/', include(api.urls)),
+    url(r'^api/', include(api.urls)), # ^api/v1/ only
     url(r'^accounts/', include(userprofile.urls)),
     url(r'^', include(redirect_old_urls.urls)),
     url(r'^', include(apps.urls)),
